Add database handler for game results
